Separated casadi solver into its own module.
